requirement.txt
asgiref==3.7.2
Django==4.2.5
django-cors-headers==4.2.0
django-filter==23.3
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==23.1
Pillow==10.0.1
psycopg2==2.9.7
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
whitenoise==6.5.0
